+++ R/HermesData-class.R
@@ -39,9 +39,9 @@ NULL
 #'   method, which allows us to turn off the validity checks in internal
 #'   functions where intermediate objects may not be valid within the scope of
 #'   the function.
-#'   - It can be helpful to convert character variables to factors in `colData()`
+#'   - It can be helpful to convert character and logical variables to factors in `colData()`
 #'   (before or after the `HermesData` creation). We provide the utility function
-#'   [df_char_to_factor()] to simplify this task, but leave it to the user to allow
+#'   [df_cols_to_factor()] to simplify this task, but leave it to the user to allow
 #'   for full control of the details.
 #'
 #' @slot prefix common prefix of the gene IDs (row names).